+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_functions_and_methods/mod.rs
@@ -1,6 +1,6 @@
 use swc_ecma_ast::{TsFnParam, TsType, TsTypeAnn};
 
-use super::GetTsType;
+use crate::ts_ast::traits::GetTsType;
 
 pub trait FunctionAndMethodTypeHelperMethods {
     fn get_ts_type_ann(&self) -> TsTypeAnn;
@@ -12,5 +12,23 @@ pub trait FunctionAndMethodTypeHelperMethods {
         self.get_ts_fn_params().iter().fold(vec![], |acc, param| {
             vec![acc, vec![param.get_ts_type().clone()]].concat()
         })
+    }
+}
+
+impl GetTsType for TsFnParam {
+    fn get_ts_type(&self) -> TsType {
+        match self {
+            TsFnParam::Ident(identifier) => match &identifier.type_ann {
+                Some(param_type) => param_type.get_ts_type(),
+                None => panic!("Function parameters must have a type"),
+            },
+            TsFnParam::Array(_) => {
+                panic!("Array destructuring in parameters is unsupported at this time")
+            }
+            TsFnParam::Rest(_) => panic!("Rest parameters are not supported at this time"),
+            TsFnParam::Object(_) => {
+                panic!("Object destructuring in parameters is unsupported at this time")
+            }
+        }
     }
 }
