--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_fn_param/get_name.rs
@@ -1,14 +1,0 @@
-use swc_ecma_ast::TsFnParam;
-
-use crate::ts_ast::GetName;
-
-impl GetName for TsFnParam {
-    fn get_name(&self) -> &str {
-        match self {
-            TsFnParam::Ident(identifier) => identifier.id.get_name(),
-            TsFnParam::Array(_) => todo!(),
-            TsFnParam::Rest(_) => todo!(),
-            TsFnParam::Object(_) => todo!(),
-        }
-    }
-}
