--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_fn_param/get_ts_type.rs
@@ -1,21 +1,0 @@
-use swc_ecma_ast::{TsFnParam, TsType};
-
-use crate::ts_ast::GetTsType;
-
-impl GetTsType for TsFnParam {
-    fn get_ts_type(&self) -> TsType {
-        match self {
-            TsFnParam::Ident(identifier) => match &identifier.type_ann {
-                Some(param_type) => param_type.get_ts_type(),
-                None => panic!("Function parameters must have a type"),
-            },
-            TsFnParam::Array(_) => {
-                panic!("Array destructuring in parameters is unsupported at this time")
-            }
-            TsFnParam::Rest(_) => panic!("Rest parameters are not supported at this time"),
-            TsFnParam::Object(_) => {
-                panic!("Object destructuring in parameters is unsupported at this time")
-            }
-        }
-    }
-}
