--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_fn_param/mod.rs
@@ -1,2 +1,0 @@
-mod get_name;
-mod get_ts_type;
